more array fixes · 8f8c1a6f qinsoon authored Dec 16, 2018
commit 8f8c1a6f7e3a8dfb5a2d6c4d2f2d0f2d4b2d1c5a
Author

+++ brepair/array.h
@@ -37,7 +37,7 @@ typedef struct
      long size;
      long fst; // first of circular array
      float factor;
-     int minsize;
+     long minsize;
      void *Rec; // records
    } Tarray;
